+++ Changes.md
@@ -1,5 +1,12 @@
 
 
+
+#
+### **+05:30 12:26:25 AM 07-09-2024, Saturday**
+
+  - `generateGerbers()` now reads the configuration `kie_include_drill` and generated the drill files first based on the value. Default value is `true`.
+  - New Version: `0.0.16`.
+
 #
 ### **+05:30 12:02:33 AM 07-09-2024, Saturday**